requirements-dev.txt
# Production dependencies
-r requirements.txt

# Code quality tools
black==24.1.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development utilities
python-dotenv==1.0.0
